Show versions filter if any shown project has versions
Previously the version filter was only shown when the current project
has versions. This is often not the case for parent projects. The filter
itself handled versions from subproject including choosing the latest
one by default but because the filter was not visible there was no
way to select "all" versions. This commit fixes the visible condition.
It is no possible to filter for "all" versions and therefore see all
issues.

Fixes #103 and fixes #96.