Resolve internal server error when grouping by parent
The filter function called an additional `all` on a resource set that
was an array already.

This CL further adds support to loading custom fixtures, tests for
grouping by parent issues and a new tooling-compatible changelog.

Note: Developers need to re-run `./redmine install` to set up a new
symlink needed for loading fixtures.

Resolves #142