feat: Support edit_own_issues permission
Users will be able to move/update their own issue when allowed using the
`:edit_own_issues` permission.

Instead of checking the `:edit_issues` permission on the project, the
controller will now use `issue.attributes_editable?` to check if a users
is allowed to update the issue. This will handle both (and future)
permissions.